requirements.txt
aerich==0.5.3
aiosqlite==0.16.1
asyncpg==0.22.0
Authlib==0.15.3
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
ddlparse==1.9.0
dictdiffer==0.8.1
fastapi==0.63.0
h11==0.12.0
iso8601==0.1.14
passlib==1.7.4
psycopg2-binary==2.8.6
pycparser==2.20
pydantic==1.8.1
PyJWT==2.1.0
pyparsing==2.4.7
pypika-tortoise==0.1.0
pytz==2020.5
six==1.15.0
starlette==0.13.6
tortoise-orm==0.17.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
